Remove /etc/sysconfig/node_exporter
This does not seem to be a necessary requirement for any platform, it is an empty file and causes errors when run on Debian 10 due to the folder `/etc/sysconfig` not existing.  The other alternative is to have puppet ensure the directory exists first but if the file is not referenced and is empty we should just remove the file declaration.